Clean out more MSVC warnings

+++ vp9/common/vp9_idct.c
@@ -18,7 +18,7 @@
 
 static INLINE uint8_t clip_pixel_add(uint8_t dest, tran_high_t trans) {
   trans = WRAPLOW(trans, 8);
-  return clip_pixel((int)WRAPLOW(dest + trans, 8));
+  return clip_pixel(WRAPLOW(dest + trans, 8));
 }
 
 void vp9_iwht4x4_16_add_c(const tran_low_t *input, uint8_t *dest, int stride) {
